Enable support for pm-proxy and fix fstab
As modem is now stable we can enable pm-proxy

fstab: Correct the metadata partition device path.

This change corrects the metadata partition's device path.
Since we need metadata to be early mounted, the bootdevice
symlink path will not work/be-available during the init's
first stage mount. This does not fails all the time but does sometimes.

Change-Id: Iccb3ec76d827d01fa33938c19c9b95f950591103