BoardConfig: Move the SMMU driver to vendor-ramdisk
Recently, the UFS driver started depending on SMMU driver via
device-tree. However, as the SMMU driver is packaged as a part
of second-stage modules, the sync-state logic disallows the ufs
driver's probe to be called in first stage. Hence, to resolve this,
move the smmu and its dependent driver to the first stage ramdisk.

Change-Id: I417a04a4a2e1eec9c9a2405dddd3252dcb094903