Rename charger_fstab.qti to charger_fw_fstab.qti for Lahaina
charger_monitor is getting enabled for non-Lahaina projects but
the component is shared with Lahaina. Since charger_fstab.qti is
already under charger_monitor project, it breaks the compilation.

Rename charger_fstab.qti to charger_fw_fstab.qti for Lahaina as
that is the quick way to address this problem.

Change-Id: I4641438ab329f40609b9c1045772b7741ecd18e1